Use UPG to clear timer counter
Simply reset() the counter register of a TIMERx seems to have
no effect on the register's content, A UPG (Update Event Generation)
needs to be trigger in order to do that, according to the datasheet
p.261 on the UPG:

This bit can be set by software, and cleared by hardware automatically. When this
bit is set, the counter is cleared if the center-aligned or up counting mode is selected,
else (down counting) it takes the auto-reload value. The prescaler counter is cleared
at the same time

To prevent generating an unwanted interrupt while triggering an UPG, we
need to set UPS to 0 so UPG does not trigger an interrupt

Additionally, UPIF should be cleared before start the timer. This is to prevent a less-common
but possible scenario where the timer is previously started (or especially, created) and
timedout (flag set) without clearing the flag, then next time when you try start() and wait()
your wait() will return immediately because of the previously set flag