Add method to get gpio port and pin information
Sometimes we need to get gpio port and pin information, for example, to
set up an external interrupt on a gpio.

`port()` and `pin_number()` are added to a gpio pin for this

In addition `pin_number()` is also added for a downgraded pin. We don't
really need a `port()` for a downgraded pin since its port information
can be obtained on the enum invariant itself