features: adds `rt` feature for gd32vf103-pac/rt
Adds a `rt` feature to allow users to opt-in to using the
`gd32vf103-pac/rt` feature. This exposes the `interrupt` macro, allowing
users to define functions for system interrupts.

This is preferable to using inline assembly to include the code in
`eclic-mode-hack.S` which causes symbol collisions with user-defined
interrupt functions, and strange bugs for handler execution.